requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airports-py==2.0.0
alembic==1.16.5
anki==24.11
anki_deck_from_text==1.3.4
annotated-types==0.7.0
appdirs==1.4.4
APScheduler==3.11.0
asgiref==3.7.2
attrs==25.3.0
beautifulsoup4==4.12.3
bitwarden_secrets_manager_python @ file:///home/batir/Pobrane/bitwarden_secrets_manager_python-1.0.0
blinker==1.9.0
cached-property==2.0.1
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.8
contourpy==1.2.0
cryptography==45.0.5
cycler==0.12.1
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
Django==5.0.2
dnspython==2.8.0
doppler-env==0.3.1
dotenv==0.9.9
email-validator==2.3.0
fernet==1.0.1
filelock==3.18.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
frozendict==2.4.0
frozenlist==1.7.0
genanki==0.13.1
google_search_results==2.4.2
greenlet==3.2.4
html5lib==1.1
idna==3.6
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
lxml==5.1.0
Mako==1.3.10
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.6.4
multitasking==0.0.11
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
orjson==3.10.14
packaging==23.2
pandas==2.2.3
peewee==3.17.1
pillow==11.0.0
platformdirs==4.3.8
plotly==5.19.0
propcache==0.3.2
protobuf==5.29.3
psycopg2-binary==2.9.10
pyaes==1.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.1.1
PyPDF2==3.0.1
pyperclip==1.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.1
pyvips==3.0.0
PyYAML==6.0.2
requests==2.32.3
scapy==2.5.0
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
shrimpy-python==0.0.15
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.43
sqlparse==0.4.4
telepot==12.7
tenacity==8.2.3
tifffile==2025.2.18
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.1
tzlocal==5.3.1
urllib3==2.2.1
virtualenv==20.31.2
webencodings==0.5.1
websockets==12.0
Werkzeug==3.1.3
yarl==1.20.1
yfinance==0.2.37
